requirements.txt
aioredis==2.0.0
aiosmtplib==1.1.6
alembic==1.7.3
amqp==5.0.6
anyio==3.3.3
asgiref==3.4.1
async-timeout==3.0.1
bcrypt==3.2.0
billiard==3.6.4.0
blinker==1.4
cached-property==1.5.2
celery==5.1.2
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.7
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==35.0.0
databases==0.5.2
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fakeredis==1.6.1
fastapi==0.68.1
fastapi-mail==1.0.1
greenlet==1.1.2
h11==0.12.0
httpcore==0.13.7
httpx==0.19.0
idna==3.2
importlib-metadata==4.8.1
importlib-resources==5.2.2
Jinja2==3.0.2
kombu==5.1.0
Mako==1.1.5
MarkupSafe==2.0.1
packaging==21.0
passlib==1.7.4
pkg_resources==0.0.0
prompt-toolkit==3.0.20
psycopg2-binary==2.8.5
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
pyparsing==2.4.7
python-decouple==3.5
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.3
redis==3.5.3
rfc3986==1.5.0
rsa==4.7.2
semantic-version==2.8.5
setuptools-rust==0.12.1
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.25
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.2
uvicorn==0.15.0
vine==5.0.0
wcwidth==0.2.5
zipp==3.6.0
